nkase-int-demo-code/dfir-api/requirements.txt
aioboto3==14.3.0
aiobotocore==2.22.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aioitertools==0.12.0
aioredis==2.0.1
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
arq==0.26.3
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
boto3==1.37.3
botocore==1.37.3
click==8.2.1
fastapi==0.115.12
frozenlist==1.6.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
hiredis==3.2.1
idna==3.10
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.4
packaging==25.0
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.3.0
rq==2.3.3
s3transfer==0.11.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wrapt==1.17.2
yarl==1.20.0
